requirements.txt
altair==5.5.0
ansi2html==1.9.1
attrs==25.4.0
blinker==1.7.0
cachetools==6.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.0
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.6
importlib-metadata==7.0.0
influxdb-client==1.49.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==2.1.3
matplotlib==3.10.7
narwhals==2.9.0
nest-asyncio==1.5.8
numpy==2.3.4
packaging==23.2
pandas==2.3.3
patsy==1.0.2
pillow==11.3.0
plotly==6.4.0
protobuf==6.33.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
reactivex==4.0.4
referencing==0.37.0
requests==2.32.5
retrying==1.3.4
rpds-py==0.28.0
scipy==1.16.3
six==1.16.0
smmap==5.0.2
statsmodels==0.14.5
streamlit==1.51.0
streamlit-plotly-events==0.0.6
tenacity==8.2.3
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.1.0
watchdog==6.0.0
Werkzeug==3.0.1
zipp==3.17.0
